requirements.txt
streamlit>=1.32,<2.0
transformers>=4.39,<5.0
torch>=2.2
scikit-learn>=1.4,<2.0
pypdf>=4.0
python-docx>=1.1
sentencepiece>=0.2
numpy>=1.26,<3.0
